feat: static verify page
